Split schema macros and decorations in modules
